+++ lib/core/di/injection_container.dart
@@ -1,11 +1,12 @@
-import 'package:blogify_flutter_main/data/data_sources/mock_comments_data_source_impl.dart';
-import 'package:blogify_flutter_main/data/data_sources/mock_posts_data_source_impl.dart';
-import 'package:blogify_flutter_main/data/data_sources/mock_users_data_source_impl.dart';
+import 'package:blogify_flutter_main/data/data_sources/remote/mock_comments_data_source_impl.dart';
+import 'package:blogify_flutter_main/data/data_sources/remote/mock_posts_data_source_impl.dart';
+import 'package:blogify_flutter_main/data/data_sources/remote/mock_users_data_source_impl.dart';
 import 'package:blogify_flutter_main/data/database/database_manager.dart';
-import 'package:blogify_flutter_main/domain/data_sources/comments_data_source.dart';
-import 'package:blogify_flutter_main/domain/data_sources/posts_data_source.dart';
-import 'package:blogify_flutter_main/domain/data_sources/users_data_source.dart';
+import 'package:blogify_flutter_main/domain/data_sources/remote/comments_data_source.dart';
+import 'package:blogify_flutter_main/domain/data_sources/remote/posts_data_source.dart';
 import 'package:get_it/get_it.dart';
+
+import '../../domain/data_sources/remote/users_data_source.dart';
 
 final serviceLocator = GetIt.instance;
