+++ lib/data/providers/global_mock_comment_provider.dart
@@ -1,4 +1,4 @@
-import 'package:blogify_flutter_main/domain/data_sources/comments_data_source.dart';
+import 'package:blogify_flutter_main/domain/data_sources/remote/comments_data_source.dart';
 import 'package:blogify_flutter_main/domain/entities/comment_entity.dart';
 import 'package:flutter/cupertino.dart';
