+++ lib/domain/data_sources/remote/users_data_source.dart
@@ -1,4 +1,4 @@
-import '../entities/user_entity.dart';
+import '../../entities/user_entity.dart';
 
 abstract class UsersDataSource {
   late List<UserEntity> users;
